HOME240513

+++ kran.h
@@ -11,14 +11,14 @@
 
 
 #endif
-const int switch_time = 7000;
+const uint16_t switch_time = 7000;
 const uint8_t quantumT=70;
 
 class Kran {
 private:
 	boolean inQuantum;
 	boolean inProgress;
-	uint16_t progress_time;
+	long progress_time;
 	uint8_t close_pin;
 	uint8_t open_pin;
 	uint8_t measure_pin;
@@ -31,7 +31,7 @@ public:
 	void forceClose();
 	void openQuantum(int8_t oq);
 	void shiftQuantum(int8_t oq);
-	void process(uint16_t ms);
+	void process(long ms);
 	boolean measureState();
 	
 	void setup(uint8_t c_pin, uint8_t o_pin, uint8_t m_pin, uint8_t r_pin);
